refactor pytorch, tests pass
