+++ galaxy_datasets/deprecated/decals_dr8.py
