+++ galaxy_datasets/prepared_datasets/__init__.py
@@ -1,0 +1,1 @@
+from galaxy_datasets.prepared_datasets import candels, decals, gz2, rings, tidal, legs, hubble
