+++ galaxy_datasets/prepared_datasets/rings.py
@@ -1,0 +1,38 @@
+import os
+
+import pandas as pd
+
+from galaxy_datasets.prepared_datasets import download_utils, internal_urls
+
+# TODO could eventually refactor this out of Zoobot as well
+from zoobot.shared import label_metadata
+
+
+
+
+def setup(root, train, download):
+    resources = [
+        (internal_urls.rings_train_catalog, 'e2fb6b2bca45cd7f1c58f5b4089a5976'),  # train catalog
+        (internal_urls.rings_test_catalog, '6e3f362a6e19ecd02675eaa48f6727f0'),  # test catalog
+        (internal_urls.rings_images, 'd0950250436a05ce88de747e6af825b6')  # the images
+    ]
+    images_to_spotcheck = []
+
+    downloader = download_utils.DatasetDownloader(root, resources, images_to_spotcheck)
+    if download is True:
+        downloader.download()
+
+    # useful_columns = label_cols + ['subfolder', 'filename']
+    useful_columns = None
+    if train:
+        train_catalog_loc = os.path.join(root, 'rings_train_catalog.parquet')
+        catalog = pd.read_parquet(train_catalog_loc, columns=useful_columns)
+    else:
+        test_catalog_loc = os.path.join(root, 'rings_test_catalog.parquet')
+        catalog = pd.read_parquet(test_catalog_loc, columns=useful_columns)
+
+    catalog['file_loc'] = catalog.apply(lambda x: os.path.join(downloader.image_dir, x['subfolder'], x['filename']), axis=1)
+
+    label_cols = label_metadata.rings_label_cols
+
+    return catalog, label_cols
